+++ app/src/main/java/com/ulan/timetable/activities/NotesActivity.java
@@ -1,6 +1,5 @@
 package com.ulan.timetable.activities;
 
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.SparseBooleanArray;
@@ -10,14 +9,14 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.AbsListView;
-import android.widget.AdapterView;
 import android.widget.ListView;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.ulan.timetable.R;
 import com.ulan.timetable.adapters.NotesAdapter;
 import com.ulan.timetable.model.Note;
-import com.ulan.timetable.R;
 import com.ulan.timetable.utils.AlertDialogsHelper;
 import com.ulan.timetable.utils.DbHelper;
 
@@ -25,8 +24,10 @@ import java.util.ArrayList;
 
 public class NotesActivity extends AppCompatActivity {
 
-    public static String KEY_NOTE = "note";
-    private Context context = this;
+    @NonNull
+    public static final String KEY_NOTE = "note";
+    @NonNull
+    private final AppCompatActivity context = this;
     private ListView listView;
     private DbHelper db;
     private NotesAdapter adapter;
@@ -49,13 +50,10 @@ public class NotesActivity extends AppCompatActivity {
         listView = findViewById(R.id.notelist);
         adapter = new NotesAdapter(NotesActivity.this, listView, R.layout.listview_notes_adapter, db.getNote());
         listView.setAdapter(adapter);
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Intent intent = new Intent(context, NoteInfoActivity.class);
-                intent.putExtra(KEY_NOTE, adapter.getNoteList().get(position));
-                startActivity(intent);
-            }
+        listView.setOnItemClickListener((parent, view, position, id) -> {
+            Intent intent = new Intent(context, NoteInfoActivity.class);
+            intent.putExtra(KEY_NOTE, adapter.getNoteList().get(position));
+            startActivity(intent);
         });
     }
 
@@ -63,14 +61,14 @@ public class NotesActivity extends AppCompatActivity {
         listView.setChoiceMode(ListView.CHOICE_MODE_MULTIPLE_MODAL);
         listView.setMultiChoiceModeListener(new AbsListView.MultiChoiceModeListener() {
             @Override
-            public void onItemCheckedStateChanged(ActionMode mode, int position, long id, boolean checked) {
+            public void onItemCheckedStateChanged(@NonNull ActionMode mode, int position, long id, boolean checked) {
                 final int checkedCount = listView.getCheckedItemCount();
                 mode.setTitle(checkedCount + " " + getResources().getString(R.string.selected));
-                if(checkedCount == 0) mode.finish();
+                if (checkedCount == 0) mode.finish();
             }
 
             @Override
-            public boolean onCreateActionMode(ActionMode mode, Menu menu) {
+            public boolean onCreateActionMode(@NonNull ActionMode mode, Menu menu) {
                 MenuInflater menuInflater = mode.getMenuInflater();
                 menuInflater.inflate(R.menu.toolbar_action_mode, menu);
                 return true;
@@ -82,30 +80,29 @@ public class NotesActivity extends AppCompatActivity {
             }
 
             @Override
-            public boolean onActionItemClicked(final ActionMode mode, MenuItem item) {
-                switch (item.getItemId()) {
-                    case R.id.action_delete:
-                        ArrayList<Note> removelist = new ArrayList<>();
-                        SparseBooleanArray checkedItems = listView.getCheckedItemPositions();
-                        for (int i = 0; i < checkedItems.size(); i++) {
-                            int key = checkedItems.keyAt(i);
-                            if (checkedItems.get(key)) {
-                                db.deleteNoteById(adapter.getItem(key));
-                                removelist.add(adapter.getNoteList().get(key));
-                            }
+            public boolean onActionItemClicked(@NonNull final ActionMode mode, @NonNull MenuItem item) {
+                if (item.getItemId() == R.id.action_delete) {
+                    ArrayList<Note> removelist = new ArrayList<>();
+                    SparseBooleanArray checkedItems = listView.getCheckedItemPositions();
+                    for (int i = 0; i < checkedItems.size(); i++) {
+                        int key = checkedItems.keyAt(i);
+                        if (checkedItems.get(key)) {
+                            db.deleteNoteById(adapter.getItem(key));
+                            removelist.add(adapter.getNoteList().get(key));
                         }
-                        adapter.getNoteList().removeAll(removelist);
-                        db.updateNote(adapter.getNote());
-                        adapter.notifyDataSetChanged();
-                        mode.finish();
-                        return true;
-
-                    default:
-                        return false;
+                    }
+                    adapter.getNoteList().removeAll(removelist);
+                    db.updateNote(adapter.getNote());
+                    adapter.notifyDataSetChanged();
+                    mode.finish();
+                    return true;
                 }
+                return false;
             }
+
             @Override
-            public void onDestroyActionMode(ActionMode mode) { }
+            public void onDestroyActionMode(ActionMode mode) {
+            }
         });
     }
